cover pages set up
