--- client/src/pages/backCover.js
@@ -1,13 +1,0 @@
-import pic from "./images/cover.png"
-
-
-function BackCover () {
-    return (
-        <div className="page-cover">
-            <h1>End YearBook</h1>
-            <img src={pic} alt="pic"/>
-        </div>
-    )
-}
-
-export default BackCover
