+++ client/src/pages/collage.js
@@ -14,6 +14,20 @@ const ClassCollage = () => {
     </div>
     );
   };
+
+  const PetCollage = () => {
+ 
+    return (
+     <div>
+        <Image 
+            width={600}
+            height={800}
+            src={require ('./images/pet-collage.JPEG')}
+            alt='pet collage'
+        />
+     </div>
+    );
+  };
   
-export default ClassCollage;
+export {ClassCollage, PetCollage};
