+++ client/src/pages/styles.scss
@@ -100,6 +100,34 @@
     }
   }
 
-  .hideP p {
-display: none;
+  .cover-content {
+    position: absolute;
+    transform: translate(-50%, -50%);
+    top: 50%;
+    left: 30%;
+    text-align: center;
+  }
+
+  .book-cover-pic {
+    width: 100%;
+    height: 100vh;
+    position: relative;
+  }
+  
+  .cover-img {
+    width: 100%;
+    height: 100%;
+    opacity: 0.75;
+    
+  }
+  
+  .cover-pic-set::after {
+    cover-content: "";
+    width: 100%;
+    height: 100%;
+    position: absolute;
+    top: 0;
+    left: 0;
+    background-color: black;
+    opacity: 0.7;
   }
